172911 - code updated AB#172911
