+++ src/UKHO.S100PermitService.Common/Encryption/IAesEncryption.cs
@@ -1,0 +1,7 @@
+﻿namespace UKHO.S100PermitService.Common.Encryption
+{
+    public interface IAesEncryption
+    {
+        string Decrypt(string hexString, string keyHexEncoded);
+    }
+}
