+++ src/UKHO.S100PermitService.Common/Encryption/IS100Crypt.cs
@@ -1,0 +1,7 @@
+﻿namespace UKHO.S100PermitService.Common.Encryption
+{
+    public interface IS100Crypt
+    {
+        string DecryptData(string hexString, string keyHexEncoded);
+    }
+}
