--- src/UKHO.S100PermitService.Common/Securities/IS100Crypt.cs
@@ -1,7 +1,0 @@
-﻿namespace UKHO.S100PermitService.Common.Securities
-{
-    public interface IS100Crypt
-    {
-        string Decrypt(string hexString, string keyHexEncoded);
-    }
-}
