--- src/UKHO.S100PermitService.Common/Securities/IS100Manufacturer.cs
@@ -1,7 +1,0 @@
-﻿namespace UKHO.S100PermitService.Common.Securities
-{
-    public interface IS100Manufacturer
-    {
-        string DecryptData(string hexString, string keyHexEncoded);
-    }
-}
