+++ src/UKHO.S100PermitService.Common/Services/PermitService.cs
@@ -5,7 +5,7 @@ using UKHO.S100PermitService.Common.IO;
 using UKHO.S100PermitService.Common.Models.Holdings;
 using UKHO.S100PermitService.Common.Models.Permits;
 using UKHO.S100PermitService.Common.Models.ProductKeyService;
-using UKHO.S100PermitService.Common.Securities;
+using UKHO.S100PermitService.Common.Encryption;
 
 namespace UKHO.S100PermitService.Common.Services
 {
@@ -18,40 +18,41 @@ namespace UKHO.S100PermitService.Common.Services
         private readonly IHoldingsService _holdingsService;
         private readonly IUserPermitService _userPermitService;
         private readonly IProductKeyService _productKeyService;
-        private readonly IS100Manufacturer _s100Manufacturer;
+        private readonly IS100Crypt _s100Crypt;
 
         public PermitService(IPermitReaderWriter permitReaderWriter,
                                 ILogger<PermitService> logger,
                                 IHoldingsService holdingsService,
                                 IUserPermitService userPermitService,
                                 IProductKeyService productKeyService,
-                                IS100Manufacturer s100Manufacturer)
+                                IS100Crypt s100Crypt)
         {
             _permitReaderWriter = permitReaderWriter ?? throw new ArgumentNullException(nameof(permitReaderWriter));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _holdingsService = holdingsService ?? throw new ArgumentNullException(nameof(holdingsService));
             _userPermitService = userPermitService ?? throw new ArgumentNullException(nameof(userPermitService));
             _productKeyService = productKeyService ?? throw new ArgumentNullException(nameof(productKeyService));
-            _s100Manufacturer = s100Manufacturer ?? throw new ArgumentNullException(nameof(s100Manufacturer));
+            _s100Crypt = s100Crypt ?? throw new ArgumentNullException(nameof(s100Crypt));
         }
 
         public async Task CreatePermitAsync(int licenceId, CancellationToken cancellationToken, string correlationId)
         {
             _logger.LogInformation(EventIds.CreatePermitStart.ToEventId(), "CreatePermit started");
 
+            var userPermitServiceResponse = await _userPermitService.GetUserPermitAsync(licenceId, cancellationToken, correlationId);
+
             var holdingsServiceResponse = await _holdingsService.GetHoldingsAsync(licenceId, cancellationToken, correlationId);
 
             var productsList = GetProductsList();
-
-            var userPermitServiceResponse = await _userPermitService.GetUserPermitAsync(licenceId, cancellationToken, correlationId);
 
             var productKeyServiceRequest = ProductKeyServiceRequest(holdingsServiceResponse);
 
             var pksResponseData = await _productKeyService.GetPermitKeysAsync(productKeyServiceRequest, cancellationToken, correlationId);
 
-            const string Upn = "ABCDEFGHIJKLMNOPQRSTUVYXYZ";
-
-            CreatePermitXml(DateTimeOffset.Now, "AB", "ABC", Upn, "1.0", productsList);
+            foreach (var userPermits in userPermitServiceResponse.UserPermits)
+            {
+                CreatePermitXml(DateTimeOffset.Now, "AB", "ABC", userPermits.Upn, "1.0", productsList);
+            };
 
             _logger.LogInformation(EventIds.CreatePermitEnd.ToEventId(), "CreatePermit completed");
         }
